fixed RPN not being able to output negative numbers

+++ module 09/ex01/incl/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: hlibine <hlibine@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/24 10:15:06 by hlibine           #+#    #+#             */
-/*   Updated: 2025/05/07 17:01:09 by hlibine          ###   ########.fr       */
+/*   Updated: 2025/05/08 13:58:13 by hlibine          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 
 class RPN {
 	public:
-		static size_t calculate(const std::string in);
+		static long calculate(const std::string in);
 		virtual ~RPN() = 0;
 };
